+++ javascripts/hotkey.js
@@ -1,14 +1,11 @@
-Mousetrap.bind("m", function() {});
-for (let tier = 0; tier < 8; tier++) {
+for (let tier = 0; tier < 3; tier++) {
 	Mousetrap.bind((tier+1).toString(), function() {
-		player.reactors[tier].buyMax();
+		player.reactors.pebblebeds[tier].buyMax();
+	});
+	Mousetrap.bind("m + " + (tier+1).toString(), function() {
+		player.reactors.pebblebeds[tier].mineFuel();
+	});
+	Mousetrap.bind("e + " + (tier+1).toString(), function() {
+		player.reactors.pebblebeds[tier].ejectWaste();
 	});
 }
-
-Mousetrap.bind("n", function(){
-	buyMaxNucleosynthesis();
-});
-
-Mousetrap.bind("r", function(){
-	buyNaniteResearch();
-});