+++ javascripts/react/dev/production/reactors/PebblebedFissionReactor.jsx
@@ -8,7 +8,6 @@ function PebblebedFissionReactorComponent(props) {
 	const [unlockedMines, setUnlockedMines] = React.useState(false);
 	const [amount, setAmount] = React.useState(zero);
 	const [bought, setBought] = React.useState(0);
-	const [canBuyFuel, setCanBuyFuel] = React.useState(false);
 	const [canLoadFuel, setCanLoadFuel] = React.useState(false);
 	const [canEjectWaste, setCanEjectWaste] = React.useState(false);
 	const [fuel, setFuel] = React.useState(zero);
@@ -24,11 +23,10 @@ function PebblebedFissionReactorComponent(props) {
 	React.useEffect(function() {
 		const timerID = setInterval(function() {
 			setUnlocked(props.tier === 0 || player.reactors.pebblebeds[props.tier - 1].bought >  0);
-			setUnlockedFuelHandling(player.energy.gt(600));
+			setUnlockedFuelHandling(unlockedFuelHandling || player.energy.gte(600));
 			setUnlockedMines(player.mines.tier > -1);
 			setAmount(player.reactors.pebblebeds[props.tier].amount);
 			setBought(player.reactors.pebblebeds[props.tier].bought);
-			setCanBuyFuel(player.reactors.pebblebeds[props.tier].canBuyFuel);
 			setCanLoadFuel(player.fuels.triso[props.tier].enriched.gte(1) && player.reactors.pebblebeds[props.tier].fuel.add(player.reactors.pebblebeds[props.tier].spent).add(1).lt(player.reactors.pebblebeds[props.tier].totalCapacity));
 			setCanEjectWaste(player.reactors.pebblebeds[props.tier].spent.gte(1));
 			setFuel(player.reactors.pebblebeds[props.tier].fuel);
@@ -45,9 +43,6 @@ function PebblebedFissionReactorComponent(props) {
 
 	function mineFuel() {
 		player.reactors.pebblebeds[props.tier].mineFuel();
-	}
-	function buyFuel() {
-		player.reactors.pebblebeds[props.tier].buyFuel();
 	}
 	function loadFuel() {
 		player.reactors.pebblebeds[props.tier].loadFuel();
@@ -89,7 +84,6 @@ function PebblebedFissionReactorComponent(props) {
 				<div className="flex-col vertical-top fuelhandling" style={{display: unlockedFuelHandling ? "" : "none"}}>
 					<div>Fuel Handling:</div>
 					<button onClick={mineFuel} className={bought > 0 ? "pebblebedbtn buy" : "pebblebedbtn locked"}>Mine Enriched {type} Pellets</button>
-					<button onClick={buyFuel} className={canBuyFuel ? "pebblebedbtn buy" : "pebblebedbtn locked"}>Buy Max Enriched {type} Pellets</button>
 					<button onClick={loadFuel} className={canLoadFuel ? "pebblebedbtn buy" : "pebblebedbtn locked"} style={{display: unlockedMines ? "" : "none"}}>Load Enriched {type} Pellets</button>
 					<button onClick={ejectWaste} className={canEjectWaste ? "pebblebedbtn buy" : "pebblebedbtn locked"}>Eject Spent {type} Pellets</button>
 				</div>